wgpu: Move run_copy_pipeline to utils
