+++ render/wgpu/src/surface/target.rs
@@ -2,8 +2,8 @@ use crate::backend::RenderTargetMode;
 use crate::buffer_pool::{AlwaysCompatible, PoolEntry, TexturePool};
 use crate::descriptors::Descriptors;
 use crate::globals::Globals;
-use crate::surface::commands::run_copy_pipeline;
 use crate::utils::create_buffer_with_data;
+use crate::utils::run_copy_pipeline;
 use crate::Transforms;
 use std::cell::OnceCell;
 use std::sync::Arc;
